+++ NEWS.md
@@ -1,3 +1,28 @@
+# Microsoft365R 2.2.0
+
+## OneDrive/SharePoint
+
+- Add a `list_shared_items()` method for the `ms_drive` class to access files and folders shared with you (#45).
+- Allow getting drives for groups, sites and teams by name. The first argument to the `get_drive()` method for these classes is now `drive_name`; to get a drive by ID, specify the argument name explicitly: `get_drive(drive_id=*)`
+- Add a `by_item` argument to the `delete_item()` method for drives and the `delete()` method for drive items (#21). This is to allow deletion of non-empty folders on SharePoint sites with data protection policies in place. Use with caution.
+
+## Outlook
+
+- Add a `search` argument to the `ms_outlook_folder$list_emails()` method. The default is to search in the from, subject and body of the emails.
+
+## Teams
+
+- Add `list_members()` and `get_member()` methods for teams and channels.
+- Add support for @mentions in Teams channel messages (#26).
+
+## Other
+
+- All `list_*` class methods now have `filter` and `n` arguments to filter the result set and cap the number of results, following the pattern in AzureGraph 1.3.0. The default values are `filter=NULL` and `n=Inf`. If `n=NULL`, an `ms_graph_pager` iterator object is returned instead to allow manual iteration over the results. Note that support for filtering in the underlying Graph API is somewhat uneven at the moment.
+- Experimental **read-only** support for plans, contributed by Roman Zenka.
+  - Add `get_plan()` and `list_plans()` methods to the `az_group` class. Note that only Microsoft 365 groups can have  plans, not any other type of group.
+  - To get the plan(s) for a site or team, call its `get_group()` method to retrieve the associated group, and then get the plan from the group.
+  - A plan has methods to retrieve tasks and buckets, as well as plan details.
+
 # Microsoft365R 2.1.0
 
 - Add support for sending and managing emails in Outlook. Use the `get_personal_outlook()` and `get_business_outlook()` client functions to access the emails in your personal account and work or school account, respectively. Functionality supported includes: